+++ src/main/java/nextstep/courses/domain/enroll/Capability.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.enroll;
 
 import java.util.Objects;
