+++ src/main/java/nextstep/courses/domain/enroll/SessionStatus.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.enroll;
 
 public enum SessionStatus {
     READY,
