+++ src/main/java/nextstep/courses/domain/image/ImageInfo.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.image;
 
 public class ImageInfo {
