+++ src/main/java/nextstep/courses/domain/image/Size.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.image;
 
 public class Size {
