+++ src/main/java/nextstep/courses/domain/tuition/Money.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.tuition;
 
 import java.util.Objects;
