+++ src/main/java/nextstep/courses/domain/tuition/Tuition.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.tuition;
 
 public class Tuition {
