+++ src/test/java/nextstep/courses/domain/FileExtensionTest.java
@@ -1,11 +1,11 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.image.FileExtension;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.junit.jupiter.api.Assertions.*;
 
 class FileExtensionTest {
