+++ src/test/java/nextstep/courses/domain/HeightTest.java
@@ -1,10 +1,10 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.image.Height;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.junit.jupiter.api.Assertions.*;
 
 class HeightTest {
