+++ src/test/java/nextstep/courses/domain/TuitionTest.java
@@ -1,5 +1,7 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.tuition.Money;
+import nextstep.courses.domain.tuition.Tuition;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
